backend/lambda/requirements.txt
# CPU-only PyTorch keeps the Lambda container image under 5 GB
# Install via: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
boto3>=1.28.0
